requirements.txt
paho-mqtt>=2.1.0
numpy>=2.3.0
python-dotenv>=1.1.0
